+++ Android/Android-Project-master/app/src/main/java/com/selma/constructions/model/Employee.java
@@ -13,7 +13,7 @@ public class Employee implements Serializable {
     private String email;
     private String phone;//
     private String imageUrl;
-    private int workHours;
+    private float workHours;
 
     public Employee() {}
 
@@ -100,11 +100,11 @@ public class Employee implements Serializable {
         this.imageUrl = imageUrl;
     }
 
-    public int getWorkHours() {
+    public float getWorkHours() {
         return workHours;
     }
 
-    public void setWorkHours(int workHours) {
+    public void setWorkHours(float workHours) {
         this.workHours = workHours;
     }
 }
